requirements-dev.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
djangorestframework==3.8.2
idna==2.6
oauthlib==2.1.0
py-trello==0.11.3
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
urllib3==1.22
